refactor: splitting endpoint and commands into their own files
